new commit

+++ mvc/controllers/AdminController.php
@@ -3,12 +3,12 @@ include_once('mvc/controllers/Interface/ActionInterface.php');
 include_once('mvc/helpers/handleFile.php');
 include_once('mvc/helpers/passwordEncryption.php');
 include_once('mvc/helpers/getToken.php');
+include_once('mvc/helpers/permission.php');
 
 class AdminController extends Controller implements ActionInterface
 {
 
     public $controller;
-    public $id;
 
     public function __construct()
     {
@@ -20,90 +20,80 @@ class AdminController extends Controller implements ActionInterface
 
         $email = $_POST['email'] ?? "";
         $password = $_POST['password'] ?? "";
-        if (isset($_POST['submit'])) {
-            //validation
-            $validation = validation([
-                'email' => $email,
-                'password' => $password
-            ]);
 
-            //check data in db
-            if ($validation) {
-                //$password = passwordEncryption($password);
-                $data = $this->model($this->controller . "Model")->checkLogin($email, $password)->fetch();
-            }
-
-            if (!empty($data['id'])) {
-                $model = $this->model('UserTokenModel');
-                $token = getToken(10);
-
-                $_SESSION['email'] = $email;
-                $_SESSION['token'] = $token;
-
-                $row_token = $model->count($email)->fetch();
-
-                if (!empty($row_token['allcount'])) {
-                    $action = $model->update($email, $token);
-                } else {
-                    $action = $model->create(array('email'=> $email, 'token'=> $token));
-                }
-
-                if($action) {
-                    setSessionAdmin('id', $data['id']);
-                    header("Location: " . DOMAIN . "Admin/search");
-                    return;
-                }
-            } else {
-                setSessionMessage('Login', 'Fail');
-            }
+        if (!isset($_POST['submit'])) {
+            $this->view($this->controller . "/login", ["email_input" => $email, "password_input" => $password]);
+            return;
         }
-
-        $this->view($this->controller . "/login", ["email_input" => $email, "password_input" => $password]);
-
-    }
-
-    public function search()
-    {
-
-        //check login SESSION
-        if(!checkSessionLogin('admin') || $this->checkToken()) {
-            header("Location: ".DOMAIN);
-        }
-        //GET condition
-        $email = $_GET["email"] ?? "";
-        $name = $_GET["name"] ?? "";
-        $condition = [
+        //validation
+        $validation = validation([
             'email' => $email,
-            'name' => $name
-        ];
-        //Model
-        $model = $this->model($this->controller . "Model");
-        $data = $model->searchData($condition,"");
-        $total_record = $model->searchData($condition,"getTotalRecord")->fetch();
-        //View
-        $this->view($this->controller . "/search", ["data" => $data, "total_record"=>$total_record]);
+            'password' => $password
+        ]);
+
+        //check data in db
+        if ($validation) {
+            //$password = passwordEncryption($password);
+            $data = $this->model($this->controller . "Model")->checkLogin($email, $password)->fetch();
+        }else {
+            $this->view($this->controller . "/login", ["email_input" => $email, "password_input" => $password]);
+            return;
+        }
+
+        if (empty($data['id'])) {
+            setSessionMessage('Login', 'Fail');
+            header("Location: " . DOMAIN . $this->controller . "/login");
+            return;
+        }
+
+        $model = $this->model('UserTokenModel');
+        $token = getToken(10);
+
+        $_SESSION['email'] = $email;
+        $_SESSION['token'] = $token;
+
+        $row_token = $model->count($email)->fetch();
+
+        if (!empty($row_token['allcount'])) {
+            $action = $model->update($email, $token);
+        } else {
+            $action = $model->create(array('email' => $email, 'token' => $token));
+        }
+
+        if ($action) {
+
+            setSessionAdmin('id', $data['id']);
+            setSessionAdmin('role_type', $data['role_type']);
+
+            //Redirect
+            header("Location: " . DOMAIN . $this->controller . "/search");
+            return;
+        }
+
     }
 
     public function create()
     {
         //check login SESSION
-        if(!checkSessionLogin('admin')) {
-            header("Location: ".DOMAIN);
+        if (!checkPermission() && $this->checkToken()) {
+            header("Location: " . DOMAIN);
         }
+
         $model = $this->model($this->controller . "Model");
 
         if (!isset($_POST['save'])) {
             $this->view($this->controller . "/create");
             return;
         }
+
         //validation
         $avatar = uploadFile();
-        $email = $model->searchData(['email'=>$_POST['email']], 'getTotalRecord')->fetch();
+        $email = $model->searchData(['email' => $_POST['email']], 'getTotalRecord')->fetch();
         $validation = validation($_POST);
 
         //Data
         if (!$validation) {
-            $this->view($this->controller . "/create",[
+            $this->view($this->controller . "/create", [
                 'avatar' => $avatar,
                 'name' => $_POST['name'],
                 'email' => $_POST['email'],
@@ -123,43 +113,76 @@ class AdminController extends Controller implements ActionInterface
             'password' => $password,
             'role_type' => $_POST['role_type']
         ];
+
         //Model
         $actioonSuccessfull = $model->create($data);
         //notice message action successfull
         if ($actioonSuccessfull) {
             setSessionActionSuccessful('Create');
         }
+
+        //Redirect
+        header("Location: " . DOMAIN . $this->controller . "/search");
+    }
+
+    public function search()
+    {
+        //check login SESSION && token
+        if (!checkPermission() || $this->checkToken()) {
+            header("Location: " . DOMAIN);
+        }
+
+        //GET condition
+        $email = $_GET["email"] ?? "";
+        $name = $_GET["name"] ?? "";
+        $condition = [
+            'email' => $email,
+            'name' => $name
+        ];
+
+        //Model
+        $model = $this->model($this->controller . "Model");
+        $data = $model->searchData($condition, "");
+        $total_record = $model->searchData($condition, "getTotalRecord")->fetch();
+
         //View
-        header("Location: ".DOMAIN."Admin/search");
+        $this->view($this->controller . "/search", ["data" => $data, "total_record" => $total_record]);
     }
 
     public function edit($id)
     {
         //check login SESSION
-        if(!checkSessionLogin('admin')) {
-            header("Location: ".DOMAIN);
+        if (!checkPermission() || $this->checkToken()) {
+            header("Location: " . DOMAIN);
         }
+
         $model = $this->model($this->controller . 'Model');
 
         if (!isset($_POST['save'])) {
-            $data = $model->findById($id, 'id')->fetch();
+            $data = $model->findByField($id, 'id')->fetch();
+            if(!empty($data)){
+                phpAlert(DATA_NOT_EXIST, $this->controller);
+                return;
+            }
             $this->view($this->controller . "/edit", [
                 'id' => $data['id'],
                 'name' => $data['name'],
                 'avatar' => $data['avatar'],
                 'email' => $data['email'],
-                'password'=>$data['password'],
-                'role_type'=>$data['role_type']
+                'password' => $data['password'],
+                'role_type' => $data['role_type']
             ]);
             return;
         }
+
         //Data
         $avatar = uploadFile();
+
         //Validate
         $validation = validation($_POST);
 
         if (!$validation) {
-            $this->view($this->controller . "/edit",[
+            $this->view($this->controller . "/edit", [
                 'id' => $id,
                 'avatar' => $avatar,
                 'name' => $_POST['name'],
@@ -180,46 +203,66 @@ class AdminController extends Controller implements ActionInterface
             'password' => $password,
             'role_type' => $_POST['role_type']
         ];
+
         //Model
         $actioonSuccessfull = $model->update($id, $data);
+
         //notice message action successfull
         if ($actioonSuccessfull) {
             setSessionActionSuccessful('Update');
         }
-        //View
-        header("Location: ".DOMAIN."Admin/search");
+
+        //Redirect
+        header("Location: " . DOMAIN . $this->controller . "/search");
     }
 
     public function delete($id)
     {
+
         //check login SESSION
-        if(!checkSessionLogin('admin')) {
-            header("Location: ".DOMAIN);
+        if (!checkPermission() || $this->checkToken()) {
+            header("Location: " . DOMAIN);
         }
+
         $model = $this->model($this->controller . 'Model');
+
         // find del_flag
-        $data = $model->findById($id, 'id')->fetch();
+        $data = $model->findByField($id, 'id')->fetch();
+        if(empty($data)){
+            phpAlert(DATA_NOT_EXIST, $this->controller);
+            return;
+        }
         // del_flag dirrection
-        if(!empty($data) && $data['del_flag'] == DELETED_OFF) {
-            $actioonSuccessfull = $model->update($id, ['del_flag'=>DELETED_ON]);
-        }else {
+        if (!empty($data) && $data['del_flag'] == DELETED_OFF) {
+            $actioonSuccessfull = $model->update($id, ['del_flag' => DELETED_ON]);
+        } else {
             removeFile($data['avatar']);
             $actioonSuccessfull = $model->deleteById($id);
         }
+
         //notice message action successfull
         if ($actioonSuccessfull) {
             setSessionActionSuccessful('Delete');
         }
-        header("Location: ".DOMAIN."Admin/search");
+        if($id==getSessionAdmin('id')) {
+            unset($_SESSION['admin']['id']);
+            $this->model('UserTokenModel')->deleteById($_SESSION['email']);
+        }
+        //Redirect
+        header("Location: " . DOMAIN . $this->controller . "/search");
     }
 
     public function logout()
     {
-        $this->model('UserTokenModel')->deleteById($_SESSION['email']);
-        $this->view($this->controller."/login", [
+        if (isset($_SESSION['email'])) {
+            $this->model('UserTokenModel')->deleteById($_SESSION['email']);
+        }
+
+        $this->view($this->controller . "/login", [
         ]);
 
         session_destroy();
+
     }
 
     public function checkToken()
@@ -227,14 +270,17 @@ class AdminController extends Controller implements ActionInterface
         if (!isset($_SESSION['email'])) {
             return 1;
         }
-        $data = $this->model('UserTokenModel')->findById($_SESSION['email'], 'email')->fetch();
+
+        $data = $this->model('UserTokenModel')->findByField($_SESSION['email'], 'email')->fetch();
         if (empty($data['token'])) {
             return 1;
         }
+
         $token = $data['token'];
         if ($_SESSION['token'] != $token) {
             return 1;
         }
+
         return 0;
     }
